+++ task/frontend/src/locales/en-US/pages.ts
@@ -1,5 +1,4 @@
 export default {
-  'pages.task.title': 'Task',
   'pages.task.content': 'View tasks',
   'pages.task.output.tip': 'Output',
   'pages.task.deleteall.title': 'Delete all completed tasks, confirm?',
@@ -7,7 +6,7 @@ export default {
   'pages.task.stop.tip': 'Stop',
   'pages.task.stop.title': 'Stop task {name}, confirm?',
 
-  'pages.script.title': 'Script',
+  'pages.script.content': 'Manage task scripts',
   'pages.script.add.title': 'Add script',
   'pages.script.view.title': 'View script',
   'pages.script.delete.title': 'Delete script {name}, confirm?',