+++ task/frontend/src/locales/zh-CN/pages.ts
@@ -1,5 +1,4 @@
 export default {
-  'pages.task.title': '任务',
   'pages.task.content': '查看任务',
   'pages.task.output.tip': '输出',
   'pages.task.deleteall.title': '删除所有已完成任务，确认？',
@@ -7,7 +6,7 @@ export default {
   'pages.task.stop.tip': '停止',
   'pages.task.stop.title': '停止任务 {name}，确认？',
 
-  'pages.script.title': '脚本',
+  'pages.script.content': '管理任务脚本',
   'pages.script.add.title': '添加脚本',
   'pages.script.view.title': '查看脚本',
   'pages.script.delete.title': '删除脚步 {name}，确认？',